pyproject.toml
[project]
name = "facecrime-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.10",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==4.6.1",
    "attrs==24.2.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi==0.115.4",
    "fastapi-cli==0.0.5",
    "frozenlist==1.4.1",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httptools==0.6.4",
    "httpx==0.27.2",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.4",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "multidict==6.1.0",
    "numpy==2.1.2",
    "orjson==3.10.10",
    "packaging==24.1",
    "pillow==10.4.0",
    "pip==24.0",
    "propcache==0.2.0",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pydantic-extra-types==2.9.0",
    "pydantic-settings==2.6.0",
    "pygments==2.18.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.12",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rich==13.9.3",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "starlette==0.41.2",
    "typer==0.12.5",
    "typing-extensions==4.12.2",
    "ujson==5.10.0",
    "urllib3==2.2.3",
    "uvicorn==0.31.1",
    "uvloop==0.21.0",
    "watchfiles==0.24.0",
    "websockets==13.1",
    "yarl==1.15.2",
]
